Add new security terms
