+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/terms/TermsModule.kt
@@ -18,16 +18,22 @@ object TermsModule {
         Backup("term_backup"),
         DisablingPin("term_disabling_pin"),
         PrivacyNotice("term_privacy_notice"),
+        WeNotHoldingYourAssets("term_we_not_holding_your_assets"),
         MonetizationDisclosure("term_monetization_disclosure"),
-        OpenSource("term_open_source");
+        TransactionsIrreversible("term_transactions_irreversible"),
+        OpenSource("term_open_source"),
+        LegalObligations("term_legal_obligations");
 
         val description: Int
             get() = when (this) {
                 Backup -> R.string.SettingsTerms_Backup
                 DisablingPin -> R.string.SettingsTerms_DisablingPin
                 PrivacyNotice -> R.string.SettingsTerms_PrivacyNotice
+                WeNotHoldingYourAssets -> R.string.SettingsTerms_WeNotHoldingYourAssets
                 MonetizationDisclosure -> R.string.SettingsTerms_MonetizationDisclosure
+                TransactionsIrreversible -> R.string.SettingsTerms_TransactionsIrreversible
                 OpenSource -> R.string.SettingsTerms_OpenSource
+                LegalObligations -> R.string.SettingsTerms_LegalObligations
             }
     }
